requirements.txt
qrcode==7.3.1
Pillow==9.3.0
PyYAML==5.4